ice-cream-factory-datapoints-extractor/pyproject.toml
[tool.poetry]
name = "ice cream factory datapoints extractor"
version = "1.0.0"
description = "A simple rest extractor for ingesting datapoints to CDF clean"
authors = ["Marie Solvik Lepoutre <marie.lepoutre@cognite.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
cognite-extractor-utils = "^2.2.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
extractor = "rest_extractor.extractor:main"
